requirements.txt
Pillow==9.1.0
black==22.3.0
mypy==0.950
pylint==2.13.8
